refactor: User Entity @Builder 제거
